fix hue shader bug

+++ lib/cocos2d-x/extensions/filters/shaders/ccFilterShader_hue_frag.h
@@ -4,7 +4,7 @@ precision mediump float;\n\
 varying vec2 v_texCoord;\n\
 \n\
 uniform sampler2D CC_Texture0;\n\
-uniform mediump float u_hueAdjust;\n\
+uniform float u_hueAdjust;\n\
 const vec4  kRGBToYPrime = vec4 (0.299, 0.587, 0.114, 0.0);\n\
 const vec4  kRGBToI     = vec4 (0.595716, -0.274453, -0.321263, 0.0);\n\
 const vec4  kRGBToQ     = vec4 (0.211456, -0.522591, 0.31135, 0.0);\n\
